Split up the group codes file into different files by type and added the text start marker to the Entity section args
