+++ Dxflib/Entities/EntityBuffer.cs
@@ -11,6 +11,7 @@
 
 using System.Collections.Generic;
 using Dxflib.IO;
+using Dxflib.IO.GroupCodes;
 
 namespace Dxflib.Entities
 {
