+++ Dxflib/Entities/LwPolyLineBuffer.cs
@@ -12,6 +12,7 @@
 using System.Collections.Generic;
 using Dxflib.Geometry;
 using Dxflib.IO;
+using Dxflib.IO.GroupCodes;
 
 namespace Dxflib.Entities
 {
