+++ Dxflib/IO/TaggedData.cs
@@ -10,6 +10,7 @@
 // ============================================================
 
 using System;
+using Dxflib.IO.GroupCodes;
 
 namespace Dxflib.IO
 {
